Move models to services
