+++ apps/api/src/controllers/__tests__/admin.test.ts
@@ -1,7 +1,7 @@
 import { faker } from "@faker-js/faker";
 import { SessionId } from "@repo/api-types/generated/api/hire_me/Session";
 import { authorisationrErrors } from "../../middleware/authorisation";
-import { AdminErrorCodes, adminModel } from "../../models/admin";
+import { AdminErrorCodes, adminModel } from "../../services/admin";
 import {
 	generateAdmin,
 	generateAdminSession,
