+++ apps/api/src/controllers/__tests__/role.test.ts
@@ -1,4 +1,4 @@
-import { roleModel } from "../../models/role";
+import { roleModel } from "../../services/role";
 import {
 	generateApplicationData,
 	generateCompany,
