+++ apps/api/src/middleware/__tests__/authorisation.test.ts
@@ -1,5 +1,5 @@
 import { faker } from "@faker-js/faker";
-import { AdminErrorCodes, adminModel } from "../../models/admin";
+import { AdminErrorCodes, adminModel } from "../../services/admin";
 import { getMockReq, getMockRes } from "../../testUtils/index";
 import { authorisationrErrors, authoriseRequest } from "../authorisation";
